# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserInfoResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"onlyLinked": {
						"type": "boolean"
					},
					"dailyNotification": {
						"type": "boolean"
					},
					"reminderNotification": {
						"type": "boolean"
					},
					"nowNotification": {
						"type": "boolean"
					},
					"icalUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"username",
					"onlyLinked",
					"dailyNotification",
					"reminderNotification",
					"nowNotification",
					"icalUrls"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"icalUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"dailyNotification": {
						"type": "boolean"
					},
					"reminderNotification": {
						"type": "boolean"
					},
					"nowNotification": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"username",
					"icalUrls",
					"dailyNotification",
					"reminderNotification",
					"nowNotification"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PasswordUpdateRequest": {
				"properties": {
					"newPassword": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					},
					"oldPassword": {
						"type": "string"
					}
				},
				"required": [
					"newPassword",
					"confirmPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TodoCreateRequest": {
				"properties": {
					"done": {
						"type": "boolean"
					},
					"subject": {
						"type": "string",
						"minLength": 1,
						"maxLength": 255
					},
					"enableDeadline": {
						"type": "boolean",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"isAllDay": {
						"type": "boolean",
						"nullable": true
					},
					"location": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"recurrenceRule": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"recurrenceException": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"startTimezone": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"endTimezone": {
						"type": "string",
						"nullable": true,
						"maxLength": 255
					},
					"startTime": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"endTime": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						],
						"nullable": true
					},
					"listId": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"subject"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListRequest": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 20
					},
					"color": {
						"type": "string",
						"pattern": "^#[0-9A-Fa-f]{6}$"
					}
				},
				"required": [
					"name",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"minLength": 3,
						"maxLength": 50
					},
					"password": {
						"type": "string",
						"minLength": 8,
						"maxLength": 50
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"properties": {
					"username": {
						"type": "string",
						"minLength": 1,
						"maxLength": 50
					},
					"email": {
						"type": "string",
						"format": "email",
						"minLength": 3,
						"maxLength": 50
					},
					"password": {
						"type": "string",
						"minLength": 8,
						"maxLength": 50
					}
				},
				"required": [
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoogleLoginRequest": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "Yendric Api",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Yendric Van Roey"
		}
	},
	"paths": {
		"/auth/user_data": {
			"get": {
				"operationId": "Info",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserInfoResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/auth/user_data/destroy": {
			"post": {
				"operationId": "Destroy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": []
			}
		},
		"/auth/user_data/update_password": {
			"post": {
				"operationId": "UpdatePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PasswordUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/todos": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Todo"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Store",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Todo"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoCreateRequest"
							}
						}
					}
				}
			}
		},
		"/todos/{todoId}": {
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Todo"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoCreateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Destroy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Todo"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lists": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"List"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Store",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"List"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListRequest"
							}
						}
					}
				}
			}
		},
		"/lists/{listId}": {
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"List"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Destroy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"List"
				],
				"security": [
					{
						"session": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/google": {
			"post": {
				"operationId": "Google",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GoogleLoginRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}